+++ lfd/utils/__init__.py
@@ -3,10 +3,47 @@ lfd/utils \n
 Utilities
 """
 
-from .alignment import *
-from .constants import *
-from .frames import *
-from .gp import *
-from .lasa import *
-from .plots import *
-from .utils import *
+from .alignment import align_demonstrations, computeP, plot_alignments, plot_keypoints
+from .constants import __root__, __version__, base_dir, exps_dir
+from .demonstrations import Demonstration, DemonstrationSet, append_progress_values
+from .frames import Frame, GlobalFrame
+from .gp import FrameRelevanceGP, LocalPolicyGP
+from .lasa import (
+    load_data,
+    load_data3,
+    load_lasa,
+    plot_curves,
+    plot_curves3,
+    plot_trajectories,
+)
+from .utils import load_dset, load_fdset, load_fdset2, prepare_data, set_seed
+
+__all__ = [
+    "__root__",
+    "__version__",
+    "align_demonstrations",
+    "alignment",
+    "append_progress_values",
+    "base_dir",
+    "computeP",
+    "Demonstration",
+    "DemonstrationSet",
+    "exps_dir",
+    "Frame",
+    "FrameRelevanceGP",
+    "GlobalFrame",
+    "load_data",
+    "load_data3",
+    "load_dset",
+    "load_fdset",
+    "load_fdset2",
+    "load_lasa",
+    "LocalPolicyGP",
+    "plot_alignments",
+    "plot_curves",
+    "plot_curves3",
+    "plot_keypoints",
+    "plot_trajectories",
+    "prepare_data",
+    "set_seed",
+]