+++ lfd/utils/lasa/__init__.py
@@ -3,5 +3,14 @@ lfd/utils/lasa \n
 LASA utilities
 """
 
-from .helpers import *
-from .lasa import *
+from .helpers import load_data, load_data3, plot_curves, plot_curves3, plot_trajectories
+from .lasa import load_lasa
+
+__all__ = [
+    "load_data",
+    "load_data3",
+    "load_lasa",
+    "plot_curves",
+    "plot_curves3",
+    "plot_trajectories",
+]